requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
autopep8==2.0.4
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dataclasses-json==0.5.8
dialogflow==1.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
google-api-core==1.34.1
google-auth==2.28.2
google-cloud-dialogflow==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.48.2
h11==0.14.0
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsons==1.6.3
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-enum==1.5.1
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
phonepe-sdk==1.1.0
proto-plus==1.23.0
protobuf==3.20.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.18.0
pydantic==2.7.3
pydantic-settings==2.3.0
pydantic_core==2.18.4
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rsa==4.9
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.36.3
tomli==2.0.1
twilio==9.0.2
typing-inspect==0.9.0
typing_extensions==4.10.0
typish==1.9.3
tzlocal==5.2
urllib3==1.26.12
uvicorn==0.29.0
wcwidth==0.2.13
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
